+++ packages/gakumas-engine/engine/StageEngine.js
@@ -9,8 +9,9 @@ import TurnManager from "./TurnManager";
 import { deepCopy } from "../utils";
 
 export default class StageEngine {
-  constructor(config) {
+  constructor(config, linkConfigs) {
     this.config = config;
+    this.linkConfigs = linkConfigs;
     this.logger = new StageLogger();
     this.cardManager = new CardManager(this);
     this.effectManager = new EffectManager(this);
@@ -18,6 +19,13 @@ export default class StageEngine {
     this.turnManager = new TurnManager(this);
     this.evaluator = new Evaluator(this);
     this.executor = new Executor(this);
+  }
+
+  getConfig(state) {
+    if (this.config.stage.type === "linkContest") {
+      return this.linkConfigs[state[S.linkPhase] || 0];
+    }
+    return this.config;
   }
 
   getInitialState(skipEffects = false) {
@@ -28,7 +36,7 @@ export default class StageEngine {
 
     // General
     state[S.cardUsesRemaining] = 0;
-    state[S.stamina] = this.config.idol.params.stamina;
+    state[S.stamina] = this.getConfig(state).idol.params.stamina;
     state[S.consumedStamina] = 0;
     state[S.genki] = 0;
     state[S.score] = 0;
@@ -50,6 +58,13 @@ export default class StageEngine {
     }
 
     return state;
+  }
+
+  changeIdol(state) {
+    state[S.linkPhase] += 1;
+    state[S.stamina] = this.getConfig(state).idol.params.stamina;
+    this.cardManager.changeIdol(state);
+    this.logger.log(state, "linkPhaseChange", { phase: state[S.linkPhase] });
   }
 
   startStage(state) {